requirements.txt
PyPDF2
python-dotenv
streamlit
openai
altair
tiktoken
protobuf
ninja
torch
sentence-transformers
sentencepiece
transformers
llama-index
llama-index-embeddings-huggingface
llama-index-llms-huggingface-api
llama-index-embeddings-instructor
llama-index-embeddings-openai
llama-index-embeddings-ollama
llama-index-llms-openai
xformers